backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.40.0
httpx==0.27.0
httpx-sse==0.4.0
aiohttp==3.11.11
psycopg[binary]>=3.2.0
pydantic>=2.8.0
pydantic-settings>=2.5.2
apscheduler==3.10.4
python-multipart==0.0.9
websockets==13.0
python-dotenv==1.0.1
mcp[cli]==1.3.0
